services/worker/requirements.txt
pydantic-settings
python-dotenv
supabase
redis
qdrant-client
google-genai
tiktoken
pdfplumber
python-docx
